use include_bytes! for target and definitions
