+++ src/serializer.rs
@@ -1,11 +1,10 @@
-use std::io::{Write, Read, BufReader, BufWriter};
+use std::io::{Write, BufWriter};
 use std::path::{Path, PathBuf};
 
 use files::File;
 use dirs::Dir;
 use errors::*;
 use helpers::Locatable;
-
 
 /// The object in charge of serializing `Files` and `Dirs` to some `io::Writer`.
 #[derive(Debug)]
@@ -29,17 +28,10 @@ impl<W> Serializer<W>
 
     fn write_file(&mut self, f: &File) -> Result<&mut Self> {
         // TODO: Use a buffered reader here for easy perf gains
-        let contents = BufReader::new(f.contents()?);
         write!(self.writer,
-               r#"File {{ path: r"{}", contents: &["#,
-               f.name().relative_to(&self.root)?.display())?;
-
-        for byte in contents.bytes() {
-            write!(self.writer, "{}, ", byte?)?;
-        }
-
-        writeln!(self.writer, "]")?;
-        writeln!(self.writer, "}}")?;
+               r#"        File {{ path: r"{}", contents: include_bytes!(concat!(env!("CARGO_MANIFEST_DIR"), r"/{}")) }}"#,
+               f.name().relative_to(&self.root)?.display(),
+               f.name().display())?;
 
         Ok(self)
     }
@@ -53,8 +45,10 @@ impl<W> Serializer<W>
     }
 
     fn write_dir(&mut self, d: &Dir) -> Result<&mut Self> {
-        write!(self.writer,
-               r#"Dir {{ path: r"{}", files: &["#,
+        writeln!(self.writer,
+               r#"Dir {{
+    path: r"{}",
+    files: &["#,
                d.path().relative_to(&self.root)?.display())?;
 
         for file in d.files() {
@@ -62,316 +56,37 @@ impl<W> Serializer<W>
             writeln!(self.writer, ",")?;
         }
 
-        write!(self.writer, "], subdirs: &[")?;
+        writeln!(self.writer, r#"    ],
+    subdirs: &["#)?;
         for dir in d.subdirs() {
             self.write_dir(dir)?;
             writeln!(self.writer, ",")?;
         }
-        writeln!(self.writer, "]")?;
-        writeln!(self.writer, "}}")?;
+        writeln!(self.writer, "    ]")?;
+        write!(self.writer, "}}")?;
 
         Ok(self)
     }
 
-    fn write_file_definition(&mut self) -> Result<&mut Self> {
-        writeln!(self.writer, "/// A single static asset.")?;
-
-        writeln!(self.writer, "#[derive(Clone, Debug, Hash, PartialEq)]")?;
-        writeln!(self.writer,
-                 "pub struct File {{
-                    pub path: &'static str,
-                    pub contents: &'static [u8]
-                }}")?;
-
-        writeln!(self.writer,
-                 "{}",
-                 "impl File {
-                     /// Get the file's path.
-                     #[inline]
-                     pub fn path(&self) -> &::std::path::Path {
-                         self.path.as_ref()
-                     }
-
-                    /// The file's name (everything after the last slash).
-                    pub fn name(&self) -> &str {
-                        self.path().file_name().unwrap().to_str().unwrap()
-                    }
-
-                    /// Get a Reader over the file's contents.
-                    pub fn as_reader(&self) -> ::std::io::Cursor<&[u8]> {
-                        ::std::io::Cursor::new(&self.contents)
-                    }
-
-                    /// The total size of this file in bytes.
-                    pub fn size(&self) -> usize {
-                        self.contents.len()
-                    }
-                }")?;
-
-        Ok(self)
-    }
-
-    fn write_dir_definition(&mut self) -> Result<&mut Self> {
-        // docs
-        writeln!(self.writer, "/// A directory embedded as a static asset.")?;
-
-        // struct definition
-        writeln!(self.writer, "#[derive(Clone, Debug, Hash, PartialEq)]")?;
-        writeln!(self.writer,
-                 "pub struct Dir {{
-                    pub path: &'static str,
-                    pub files: &'static [File],
-                    pub subdirs: &'static [Dir]
-                  }}")?;
-
-        // method impls
-        writeln!(self.writer,
-                 "{}",
-                 r#"
-        impl Dir {
-            /// Find a file which *exactly* matches the provided name.
-            pub fn find(&'static self, name: &str) -> Option<&'static File> {
-                for file in self.files {
-                    if file.name() == name {
-                        return Some(file);
-                    }
-                }
-
-                for dir in self.subdirs {
-                    if let Some(f) = dir.find(name) {
-                        return Some(f);
-                    }
-                }
-
-                None
-            }
-
-            /// Recursively walk the various sub-directories and files inside
-            /// the bundled asset.
-            ///
-            /// # Examples
-            ///
-            /// ```rust,ignore
-            /// for entry in ASSET.walk() {
-            ///   match entry {
-            ///     DirEntry::File(f) => println!("{} ({} bytes)",
-            ///                                   f.path().display(),
-            ///                                   f.contents.len()),
-            ///     DirEntry::Dir(d) => println!("{} (files: {}, subdirs: {})",
-            ///                                  d.path().display(),
-            ///                                  d.files.len(),
-            ///                                  d.subdirs.len()),
-            ///   }
-            /// }
-            /// ```
-            pub fn walk<'a>(&'a self) -> DirWalker<'a>
-            {
-                DirWalker::new(self)
-            }
-
-            /// Get the directory's name.
-            pub fn name(&self) -> &str {
-                self.path().file_name().map(|s| s.to_str().unwrap()).unwrap_or("")
-            }
-
-            /// The directory's full path relative to the root.
-            pub fn path(&self) -> &::std::path::Path {
-                self.path.as_ref()
-            }
-
-            /// Get the total size of this directory and its contents in bytes.
-            pub fn size(&self) -> usize {
-                let file_size = self.files.iter().map(|f| f.size()).sum();
-                
-                self.subdirs.iter().fold(file_size, |acc, d| acc + d.size())
-            }
-        }"#)?;
-
-        Ok(self)
-    }
-
-    fn write_direntry_definition(&mut self) -> Result<&mut Self> {
-        writeln!(self.writer, "{}", "/// A directory entry.")?;
-
-        writeln!(self.writer, "#[derive(Debug, PartialEq, Clone)]")?;
-        writeln!(self.writer,
-                 "{}",
-                 "pub enum DirEntry<'a> {
-                     Dir(&'a Dir),
-                     File(&'a File),
-                 }")?;
-
-        writeln!(self.writer,
-                 "{}",
-                 "impl<'a> DirEntry<'a> {
-                     /// Get the entry's name.
-                     pub fn name(&self) -> &str {
-                         match *self {
-                             DirEntry::Dir(d) => d.name(),
-                             DirEntry::File(f) => f.name(),
-                         }
-                     }
-                     
-                     /// Get the entry's path relative to the root directory.
-                     pub fn path(&self) -> &::std::path::Path {
-                         match *self {
-                             DirEntry::Dir(d) => d.path(),
-                             DirEntry::File(f) => f.path(),
-                         }
-                     }
-                 }")?;
-
-        Ok(self)
-    }
-
-    fn write_dirwalker_definition(&mut self) -> Result<&mut Self> {
-        writeln!(self.writer,
-                 "{}",
-                 "/// A directory walker.
-                  ///
-                  /// `DirWalker` is an iterator which will recursively traverse
-                  /// the embedded directory, allowing you to inspect each item.
-                  /// It is largely modelled on the API used by the `walkdir`
-                  /// crate.
-                  ///
-                  /// You probably won't create one of these directly, instead
-                  /// prefer to use the `Dir::walk()` method.")?;
-
-        writeln!(self.writer, "#[derive(Debug, PartialEq, Clone)]")?;
-        writeln!(self.writer,
-                 "{}",
-                 "pub struct DirWalker<'a> {
-                     root: &'a Dir,
-                     entries_to_visit: ::std::collections::VecDeque<DirEntry<'a>>,
-                 }")?;
-
-        writeln!(self.writer,
-                 "{}",
-                 "impl<'a> DirWalker<'a> {
-                     fn new(root: &'a Dir) -> DirWalker<'a> {
-                         let mut walker = DirWalker{
-                            root: root,
-                            entries_to_visit: ::std::collections::VecDeque::new(),
-                         };
-                         walker.extend_contents(root);
-                         walker
-                     }
-
-                     fn extend_contents(&mut self, from: &Dir) {
-                         for file in from.files {
-                             self.entries_to_visit.push_back(DirEntry::File(file));
-                         }
-
-                         for dir in from.subdirs {
-                             self.entries_to_visit.push_back(DirEntry::Dir(dir));
-                         }
-                     }
-                 }")?;
-
-        writeln!(self.writer,
-                 "{}",
-                 "impl<'a> Iterator for DirWalker<'a> {
-                    type Item = DirEntry<'a>;
-
-                    fn next(&mut self) -> Option<Self::Item> {
-                        let entry = self.entries_to_visit.pop_front();
-
-                        if let Some(DirEntry::Dir(d)) = entry {
-                            self.extend_contents(d);
-                            Some(DirEntry::Dir(d))
-                        } else {
-                            entry
-                        }
-                    }
-                }")?;
-
+    fn write_std_definitions(&mut self) -> Result<&mut Self> {
+        self.writer.write(include_bytes!("serialized_std_definitions.rs"))?;
         Ok(self)
     }
 
     #[cfg(feature = "globs")]
-    fn write_globs(&mut self) -> Result<&mut Self> {
-        writeln!(self.writer,
-                 "{}",
-                 "
-                /// An iterator over all `DirEntries` which match the specified
-                /// pattern.
-                ///
-                /// # Note
-                /// 
-                /// You probably don't want to use this directly. Instead, you'll
-                /// want the [`Dir::glob()`] method.
-                /// 
-                /// [`Dir::glob()`]: struct.Dir.html#method.glob
-                pub struct Globs<'a> {
-                    walker: DirWalker<'a>,
-                    pattern: ::glob::Pattern,
-                }")?;
-        writeln!(self.writer,
-                 "{}",
-                 r#"
-                 impl<'a> Iterator for Globs<'a> {
-                    type Item = DirEntry<'a>;
-                    fn next(&mut self) -> Option<Self::Item> {
-                        while let Some(entry) = self.walker.next() {
-                            if self.pattern.matches_path(entry.path()) {
-                                return Some(entry);
-                            }
-                        }
-
-                        None
-                    }
-                 }
-        "#)?;
-
-        writeln!(self.writer,
-                 "{}",
-                 r#"impl Dir {
-                    /// Find all `DirEntries` which match a glob pattern.
-                    ///
-                    /// # Note
-                    /// 
-                    /// This may fail if you pass in an invalid glob pattern,
-                    /// consult the [glob docs] for more info on what a valid
-                    /// pattern is.
-                    ///
-                    /// # Examples
-                    ///
-                    /// ```rust,ignore
-                    /// use handlebars::Handlebars;
-                    /// let mut handlebars = Handlebars::new();
-                    ///
-                    /// for entry in ASSETS.glob("*.hbs")? {
-                    ///     if let DirEntry::File(f) = entry {
-                    ///         let template_string = String::from_utf8(f.contents.to_vec())?;
-                    ///         handlebars.register_template_string(f.name(),template_string)?;
-                    ///     }
-                    /// }
-                    /// ```
-                    /// 
-                    /// [glob docs]: https://doc.rust-lang.org/glob/glob/struct.Pattern.html
-                    pub fn glob<'a>(&'a self, pattern: &str) -> Result<Globs<'a>, Box<::std::error::Error>> {
-                        let pattern = ::glob::Pattern::new(pattern)?;
-                        Ok(Globs {
-                            walker: self.walk(),
-                            pattern: pattern,
-                        })
-                    }
-            }"#)?;
-
+    fn write_globs_definitions(&mut self) -> Result<&mut Self> {
+        self.writer.write(include_bytes!("serialized_globs_definitions.rs"))?;        
         Ok(self)
     }
 
     #[cfg(not(feature = "globs"))]
-    fn write_globs(&mut self) -> Result<&mut Self> {
+    fn write_globs_definitions(&mut self) -> Result<&mut Self> {
         Ok(self)
     }
 
     pub fn write_definitions(&mut self) -> Result<&mut Self> {
-        self.write_file_definition()?
-            .write_dir_definition()?
-            .write_dirwalker_definition()?
-            .write_globs()?
-            .write_direntry_definition()?;
+        self.write_std_definitions()?
+            .write_globs_definitions()?;
 
         Ok(self)
     }
@@ -468,33 +183,6 @@ mod tests {
         root
     }
 
-    #[test]
-    fn serialize_file_definition() {
-        let mut writer = Vec::new();
-
-        {
-            let mut serializer = Serializer::new("", &mut writer);
-            serializer.write_file_definition().unwrap();
-        }
-
-        let got = String::from_utf8(writer).unwrap();
-        assert!(got.contains("pub struct File {"));
-    }
-
-    #[test]
-    fn serialize_dir_definition() {
-        let mut writer = Vec::new();
-
-        {
-            let mut serializer = Serializer::new("", &mut writer);
-            serializer.write_dir_definition().unwrap();
-        }
-
-        let got = String::from_utf8(writer).unwrap();
-        assert!(got.contains("pub struct Dir {"));
-    }
-
-    compile_and_test!(compile_file_definition, |ser| ser.write_file_definition()?);
     compile_and_test!(compile_all_definitions, |ser| ser.write_definitions()?);
